backend/requirements.txt
fastapi[all]
uvicorn
autogen
python-dotenv
langchainhub
langchain
langgraph
langchain-google-vertexai